로그모듈 수정, 참조경로 변경 (#62)
## log 모듈 수정 
- file 로그를 사용하는 경우 30초마다 file buffer 를 flush 하게 함

## 참조 경로 수정
- `_MyLib` 의 상위경로를 기준으로 참조경로를 변경
- Monster 등에서 `$(SolutionDir)` 기준으로 _MyLib 참조를 하기때문에 _MyLib_test.vcxproj 도 그에 맞게 수정

+++ src/Win32Utils.cpp
@@ -4360,7 +4360,7 @@ std::string generate_random_string(_In_ const size_t length)
 		"{|}~";
 
 	std::stringstream strm;
-	for (int i = 0; i < length; ++i) 
+	for (size_t i = 0; i < length; ++i) 
 	{
 		strm << alphanum[rand() % (sizeof(alphanum) - 1)];
 	}
@@ -4380,7 +4380,7 @@ std::wstring generate_random_stringw(_In_ const size_t length)
 		L"{|}~";
 
 	std::wstringstream strm;
-	for (int i = 0; i < length; ++i)
+	for (size_t i = 0; i < length; ++i)
 	{
 		wchar_t w = alphanumw[rand() % ((sizeof(alphanumw) / sizeof(wchar_t)) - 1)];
 		strm << w;
